+++ cmd/create.go
@@ -59,7 +59,7 @@ var cmdCreateDemoA8s = &cobra.Command{
 	Short: "Create a demo environment for the pod based a8s Data Services such as a8s Postgres.",
 	Long:  `The demo assistent helps with the creation of a local Kubernetes cluster, installing the a8s Data Service operator(s) including necessary dependencies.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		ExecuteA8sPGDemo()
+		CreateA8sDemoEnvironment()
 	},
 }